refactor(THU-321): move mostRecentDate to shared utils
- Address PR review feedback: utility belongs in utils.ts with other
  date helpers, not in the hook file
- Move tests to utils.test.ts alongside existing utility tests
