+++ src/lib/utils.test.ts
@@ -2,7 +2,14 @@ import type { UIMessage } from 'ai'
 import { describe, expect, it } from 'bun:test'
 import { integer, sqliteTable, text } from 'drizzle-orm/sqlite-core'
 import { v7 as uuidv7 } from 'uuid'
-import { clearNullableColumns, convertUIMessageToDbChatMessage, formatNumber, hashValues, splitPartType } from './utils'
+import {
+  clearNullableColumns,
+  convertUIMessageToDbChatMessage,
+  formatNumber,
+  hashValues,
+  mostRecentDate,
+  splitPartType,
+} from './utils'
 
 describe('utils', () => {
   describe('formatNumber', () => {
@@ -183,6 +190,40 @@ describe('utils', () => {
       const hash3 = hashValues([3, 2, 1])
       expect(hash1).toBe(hash2)
       expect(hash1).not.toBe(hash3)
+    })
+  })
+
+  describe('mostRecentDate', () => {
+    it('returns b when a is null', () => {
+      const b = new Date('2024-01-02')
+      expect(mostRecentDate(null, b)).toBe(b)
+    })
+
+    it('returns a when b is null', () => {
+      const a = new Date('2024-01-02')
+      expect(mostRecentDate(a, null)).toBe(a)
+    })
+
+    it('returns null when both are null', () => {
+      expect(mostRecentDate(null, null)).toBeNull()
+    })
+
+    it('returns the more recent date when a is later', () => {
+      const a = new Date('2024-01-03')
+      const b = new Date('2024-01-01')
+      expect(mostRecentDate(a, b)).toBe(a)
+    })
+
+    it('returns the more recent date when b is later', () => {
+      const a = new Date('2024-01-01')
+      const b = new Date('2024-01-03')
+      expect(mostRecentDate(a, b)).toBe(b)
+    })
+
+    it('returns a when both dates are equal', () => {
+      const a = new Date('2024-01-01')
+      const b = new Date('2024-01-01')
+      expect(mostRecentDate(a, b)).toBe(a)
     })
   })
